updating auth for products brands and grocery items

+++ src/controllers/brand.controller.js
@@ -1,5 +1,6 @@
 import models from '../models/index.js';
 import { genericUpdateSuccessResponse } from '../utils/utils.js';
+import { isAdmin, isShopper } from '../utils/utils';
 
 const { Brand } = models;
 
@@ -26,7 +27,9 @@ const getBrandById = async (req, res, next) => {
 };
 
 const updateBrand = async (req, res, next) => {
-	if (!Brand.hasRequiredFields(req.body)) {
+	if (!isAdmin(req)) {
+		res.sendStatus(403);
+	} else if (!Brand.hasRequiredFields(req.body)) {
 		res.sendStatus(400);
 	} else {
 		try {
@@ -42,7 +45,9 @@ const updateBrand = async (req, res, next) => {
 };
 
 const createBrand = async (req, res, next) => {
-	if (!Brand.hasRequiredFields(req.body)) {
+	if (!isShopper(req)) {
+		res.sendStatus(403);
+	} else if (!Brand.hasRequiredFields(req.body)) {
 		res.sendStatus(400);
 	} else {
 		try {
@@ -55,11 +60,15 @@ const createBrand = async (req, res, next) => {
 };
 
 const deleteBrand = async (req, res, next) => {
-	try {
-		await Brand.destroy({ where: { uuid: req.params.brandId } });
-		res.sendStatus(200);
-	} catch (err) {
-		next(err);
+	if (!isAdmin(req)) {
+		res.sendStatus(403);
+	} else {
+		try {
+			await Brand.destroy({ where: { uuid: req.params.brandId } });
+			res.sendStatus(200);
+		} catch (err) {
+			next(err);
+		}
 	}
 };
