+++ src/controllers/product.controller.js
@@ -1,5 +1,6 @@
 import models from '../models/index.js';
 import { genericUpdateSuccessResponse } from '../utils/utils.js';
+import { isAdmin, isShopper } from '../utils/utils';
 
 const { Product } = models;
 
@@ -26,7 +27,9 @@ const getProductById = async (req, res, next) => {
 };
 
 const updateProduct = async (req, res, next) => {
-	if (!Product.hasRequiredFields(req.body)) {
+	if (!isAdmin(req)) {
+		res.sendStatus(403);
+	} else if (!Product.hasRequiredFields(req.body)) {
 		res.sendStatus(400);
 	} else {
 		try {
@@ -42,7 +45,9 @@ const updateProduct = async (req, res, next) => {
 };
 
 const createProduct = async (req, res, next) => {
-	if (!Product.hasRequiredFields(req.body)) {
+	if (!isShopper(req)) {
+		res.sendStatus(403);
+	} else if (!Product.hasRequiredFields(req.body)) {
 		res.sendStatus(400);
 	} else {
 		try {
@@ -55,11 +60,15 @@ const createProduct = async (req, res, next) => {
 };
 
 const deleteProduct = async (req, res, next) => {
-	try {
-		await Product.destroy({ where: { uuid: req.params.productId } });
-		res.sendStatus(200);
-	} catch (err) {
-		next(err);
+	if (!isAdmin(req)) {
+		res.sendStatus(403);
+	} else {
+		try {
+			await Product.destroy({ where: { uuid: req.params.productId } });
+			res.sendStatus(200);
+		} catch (err) {
+			next(err);
+		}
 	}
 };
